requirements.txt
cffi==1.14.5
cryptography==3.4.7
docopt==0.6.2
numpy==1.20.1
Pillow==8.1.2
pycparser==2.20
